split numeric basic arithmetic and other bit arithmetic
