+++ src/query/functions/Cargo.toml
@@ -27,11 +27,11 @@ databend-common-io = { workspace = true }
 databend-common-openai = { workspace = true }
 databend-common-vector = { workspace = true }
 databend-functions-scalar-arithmetic = { workspace = true }
-databend-functions-scalar-arithmetic-modulo = { workspace = true }
 databend-functions-scalar-datetime = { workspace = true }
 databend-functions-scalar-decimal-utils = { workspace = true }
 databend-functions-scalar-geo = { workspace = true }
 databend-functions-scalar-math = { workspace = true }
+databend-functions-scalar-numeric-basic-arithmetic = { workspace = true }
 ethnum = { workspace = true }
 geo = { workspace = true }
 geohash = { workspace = true }
