+++ src/query/functions/src/scalars/arithmetic/Cargo.toml
@@ -5,13 +5,11 @@ edition = "2021"
 
 [dependencies]
 databend-common-expression = { workspace = true }
-databend-functions-scalar-arithmetic-modulo = { workspace = true }
 databend-functions-scalar-decimal-utils = { workspace = true }
 ethnum = { workspace = true }
 lexical-core = { workspace = true }
-num-traits = { workspace = true }
-
 match-template = { workspace = true }
+num-traits = { workspace = true }
 
 [package.metadata.cargo-machete]
 ignored = ["match-template"]
