--- src/query/functions/src/scalars/arithmetic_modulo/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "databend-functions-scalar-arithmetic-modulo"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-databend-common-expression = { workspace = true }
-num-traits = { workspace = true }
-strength_reduce = { workspace = true }
