+++ src/query/functions/src/scalars/numeric_basic_arithmetic/Cargo.toml
@@ -1,14 +1,14 @@
 [package]
-name = "databend-functions-scalar-decimal-utils"
+name = "databend-functions-scalar-numeric-basic-arithmetic"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-databend-common-base = { workspace = true }
 databend-common-expression = { workspace = true }
-ethnum = { workspace = true }
+databend-functions-scalar-decimal-utils = { workspace = true }
 match-template = { workspace = true }
 num-traits = { workspace = true }
+strength_reduce = { workspace = true }
 
 [package.metadata.cargo-machete]
 ignored = ["match-template"]
